Chapter04/saas_app/requirements.txt
fastapi[all]
sqlalchemy
passlib
bcrypt==3.2.2
python-jose[cryptography]

# testing packages
pytest